Refactor and abstract questions routes to use API Config
